refactor: api reanmed services
